a bit of refactorization
