+++ examples/tensorflow/license-plate-reader/client/config.json
@@ -1,0 +1,5 @@
+{
+    "camera": {
+        
+    }
+}
