+++ examples/tensorflow/license-plate-reader/client/utils/image.py
@@ -1,0 +1,34 @@
+import cv2
+import numpy as np
+
+def resize_image(image, desired_width):
+        current_width = image.shape[1]
+        scale_percent = desired_width / current_width
+        width = int(image.shape[1] * scale_percent)
+        height = int(image.shape[0] * scale_percent)
+        resized = cv2.resize(image, (width, height), interpolation = cv2.INTER_AREA)
+        return resized
+
+def compress_image(image, grayscale=True, desired_width=416, top_crop_percent=0.45):
+    if grayscale:
+        image = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)
+    image = resize_image(image, desired_width)
+    height = image.shape[0]
+    if top_crop_percent:
+        image[:int(height * top_crop_percent)] = 128
+
+    return image
+
+def image_from_bytes(byte_im):
+    nparr = np.frombuffer(byte_im, np.uint8)
+    img_np = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
+    return img_np
+
+def image_to_jpeg_nparray(image, quality=[int(cv2.IMWRITE_JPEG_QUALITY), 95]):
+    is_success, im_buf_arr = cv2.imencode(".jpg", image, quality)
+    return im_buf_arr
+
+def image_to_jpeg_bytes(image, quality=[int(cv2.IMWRITE_JPEG_QUALITY), 95]):
+    buf = image_to_jpeg_nparray(image, quality)
+    byte_im = buf.tobytes()
+    return byte_im
